+++ Assets/Hoverboard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d629c6d7f5e27be4d9dfc9d273f7ec62
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
