+++ Assets/Scenes/SampleScene.unity
@@ -1550,7 +1550,7 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 1999818337}
+      - m_Target: {fileID: 0}
         m_TargetAssemblyTypeName: UnityEngine.GameObject, UnityEngine
         m_MethodName: SetActive
         m_Mode: 6
@@ -2484,7 +2484,7 @@ MonoBehaviour:
   m_Priority: 10
   m_StandbyUpdate: 2
   m_LookAt: {fileID: 0}
-  m_Follow: {fileID: 1523213444}
+  m_Follow: {fileID: 0}
   m_Lens:
     FieldOfView: 80
     OrthographicSize: 10
@@ -2520,11 +2520,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!4 &1635202339 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 2505484423890292237, guid: 4c5d91a383abfeb4f885b972fc2bce7f, type: 3}
-  m_PrefabInstance: {fileID: 1523213443}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1 &1665164500
 GameObject:
   m_ObjectHideFlags: 0
@@ -3393,7 +3388,7 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 1
         m_CallState: 2
-      - m_Target: {fileID: 508873713}
+      - m_Target: {fileID: 1999818349}
         m_TargetAssemblyTypeName: UnityEngine.GameObject, UnityEngine
         m_MethodName: SetActive
         m_Mode: 6
@@ -3651,7 +3646,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
-  - {fileID: 1523213444}
+  - {fileID: 2057471424}
   - {fileID: 1952027927}
   - {fileID: 1516245463}
   - {fileID: 576032223}
