+++ Assets/prefabs/HoverBoard.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2bece56881aa4c24bb7ae1ad95758083
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
